Clean up radial menu code a bit
