--- Assets/Prefabs/Individual Icon.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 48d03c1a3bfdc432385904bcb4528d8d
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
