+++ Assets/Prefabs/Models/kirby.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 17a3172f10730b146b3ff1b7d8c6825c
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
