+++ Assets/Prefabs/Models/rock.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 97fb6530e0ca14db08647c38633f5e2d
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
