--- Assets/Prefabs/RadialMenuItem.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: a301e1d14b3453745832e75f2af0e288
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
