--- Assets/Prefabs/pot.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: de1c0648423088c4bb2a1138aa55827e
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
