--- Assets/Prefabs/shrub.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: a7b0a6a0026528b4d91d535655074302
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
