--- Assets/Prefabs/tree.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 1eeab9d86a3005b47a4e087aac4ae0c3
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
