+++ Assets/Scenes/SampleScene.unity
@@ -1320,7 +1320,7 @@ GameObject:
   - component: {fileID: 1878855430}
   - component: {fileID: 1878855429}
   m_Layer: 0
-  m_Name: Word Canvas
+  m_Name: RadialMenu Canvas
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -1354,7 +1354,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 1878855427}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 60e07be32c3bd4313ad23532385a3794, type: 3}
+  m_Script: {fileID: 11500000, guid: c631580acab5e164dad110db3755174c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   radialMenuItemPrefab: {fileID: 8843825906572956133, guid: a301e1d14b3453745832e75f2af0e288,
@@ -1584,80 +1584,6 @@ Transform:
     type: 3}
   m_PrefabInstance: {fileID: 5698529633401342989}
   m_PrefabAsset: {fileID: 0}
---- !u!1001 &4600781501012911002
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 2844459549355416819, guid: a301e1d14b3453745832e75f2af0e288,
-        type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 388.5
-      objectReference: {fileID: 0}
-    - target: {fileID: 2844459549355416819, guid: a301e1d14b3453745832e75f2af0e288,
-        type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 218.5
-      objectReference: {fileID: 0}
-    - target: {fileID: 2844459549355416819, guid: a301e1d14b3453745832e75f2af0e288,
-        type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 2844459549355416819, guid: a301e1d14b3453745832e75f2af0e288,
-        type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 2844459549355416819, guid: a301e1d14b3453745832e75f2af0e288,
-        type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 2844459549355416819, guid: a301e1d14b3453745832e75f2af0e288,
-        type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 2844459549355416819, guid: a301e1d14b3453745832e75f2af0e288,
-        type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 2844459549355416819, guid: a301e1d14b3453745832e75f2af0e288,
-        type: 3}
-      propertyPath: m_RootOrder
-      value: 7
-      objectReference: {fileID: 0}
-    - target: {fileID: 2844459549355416819, guid: a301e1d14b3453745832e75f2af0e288,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 2844459549355416819, guid: a301e1d14b3453745832e75f2af0e288,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 2844459549355416819, guid: a301e1d14b3453745832e75f2af0e288,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7389209717798565771, guid: a301e1d14b3453745832e75f2af0e288,
-        type: 3}
-      propertyPath: m_Name
-      value: GameObject
-      objectReference: {fileID: 0}
-    - target: {fileID: 8888697254670273918, guid: a301e1d14b3453745832e75f2af0e288,
-        type: 3}
-      propertyPath: m_RootOrder
-      value: 7
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: a301e1d14b3453745832e75f2af0e288, type: 3}
 --- !u!4 &5006512735259261374 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 5369545388344933633, guid: 1eeab9d86a3005b47a4e087aac4ae0c3,
