+++ Assets/Scripts/UI/Icons.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 25dc1129ebf0ad74c929a321a93b870d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
