+++ Assets/Scripts/UI/RadialMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c631580acab5e164dad110db3755174c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
