--- Assets/Scripts/UI/WordSelectMenu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 60e07be32c3bd4313ad23532385a3794
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
