+++ Library/StateCache/Hierarchy/0ebc8e-mainStage.json
@@ -1,1 +1,1 @@
-﻿{"m_ExpandedPrefabGameObjectFileIDs":[],"m_ExpandedSceneGameObjectInstanceIDs":[-1168,13180,13282],"m_ScrollY":0.0,"m_LastClickedFileID":0,"m_LastClickedInstanceID":0}
+﻿{"m_ExpandedPrefabGameObjectFileIDs":[],"m_ExpandedSceneGameObjectInstanceIDs":[-1168,13180],"m_ScrollY":0.0,"m_LastClickedFileID":0,"m_LastClickedInstanceID":13282}
